Redesigned python tutorial for workflow versioning
