+++ tutorials/workflow/python/versioning/versioning.http
@@ -1,18 +1,11 @@
 @apphost=http://localhost:5263
 
-### Start the versioning workflows
+### Start the workflow
 # @name startWorkflowRequest
 POST {{ apphost }}/start
 
 
-@instanceId={{startWorkflowRequest.response.body.workflows_instances[0].instance_id}}
+@instanceId={{startWorkflowRequest.response.body.instance_id}}
 @daprHost=http://localhost:3561
-### Get the patching workflow status
+### Get the workflow status
 GET {{ daprHost }}/v1.0/workflows/dapr/{{ instanceId }}
-
-
-@instanceId={{startWorkflowRequest.response.body.workflows_instances[1].instance_id}}
-@daprHost=http://localhost:3561
-### Get the named versioned workflow status
-GET {{ daprHost }}/v1.0/workflows/dapr/{{ instanceId }}
-
