--- tutorials/workflow/python/versioning/versioning/runtime.py
@@ -1,3 +1,0 @@
-import dapr.ext.workflow as wf
-
-wf_runtime = wf.WorkflowRuntime()
